requirements.txt
pycryptodome
mysql-connector-python
pyperclip
rich
